Fix #2314: Fix BigDecimal.valueExact fast path overflows.

+++ javalib/src/main/scala/java/math/BigDecimal.scala
@@ -1728,7 +1728,7 @@ class BigDecimal() extends Number with Comparable[BigDecimal] {
    */
   private def valueExact(bitLengthOfType: Int): Long = {
     // Fast path to avoid some large BigInteger creations by toBigIntegerExact
-    if (-scale + approxPrecision() > 19) {
+    if (-scale.toLong + approxPrecision() > 19) {
       /* If there are more digits than the number of digits of Long.MaxValue in
        * base 10, this BigDecimal cannot possibly be an exact Long.
        */
